planner: move logical lock into logicalop pkg. (#55160)
ref pingcap/tidb#51664, ref pingcap/tidb#52714
